requirements.txt
flask
pymongo
psutil
python-dotenv
flask-socketio==5.3.6
python-engineio==4.5.1
python-socketio==5.8.0